#101 #94 fixing types for the contract actions
