+++ src/models/contractAppContextTypes.ts
@@ -1,6 +1,6 @@
 import { BaseContract } from 'ethers';
 
-import { TEthersAdaptor, TConnectorList } from '~~/models';
+import { TConnectorList } from '~~/models';
 
 export type TContractsByName<GContractNames extends string> = {
   [contractName in GContractNames]: { [chainId: number]: BaseContract };
@@ -8,43 +8,6 @@ export type TContractsByName<GContractNames extends string> = {
 export type TContractsByChainId<GContractNames extends string> = {
   [chainId: number]: { [contractName in GContractNames]: BaseContract };
 };
-
-export const sortContractsByChainId = <GContractNames extends string>(
-  contractsByName: TContractsByName<GContractNames>
-): TContractsByChainId<GContractNames> => {
-  const contractsByChainId: TContractsByChainId<GContractNames> = {};
-
-  Object.keys(contractsByName).forEach((name: string) => {
-    const contractName = name as GContractNames;
-    Object.keys(contractsByName[contractName])
-      .map(Number)
-      .forEach((chainId: number) => {
-        contractsByChainId[chainId][contractName] = contractsByName[contractName][chainId];
-      });
-  });
-
-  return contractsByChainId;
-};
-
-export type TContractsActionTypes<GContractNames extends string> =
-  | {
-      type: 'CONNECT_TO_CONTRACT';
-      payload: {
-        contractName: GContractNames;
-        ethersAdaptor: TEthersAdaptor;
-      };
-    }
-  | {
-      type: 'CONNECT_TO_ALL_CONTRACT';
-      payload: {
-        appContractConnectorList: TConnectorList<GContractNames>;
-        ethersAdaptor: TEthersAdaptor | undefined;
-      };
-    };
-
-export type TContractsContextActions<GContractNames extends string> = (
-  actions: TContractsActionTypes<GContractNames>
-) => Promise<void>;
 
 /**
  *
@@ -54,7 +17,6 @@ export type TContractsAppContext<GContractNames extends string> = {
   contractsByName: TContractsByName<GContractNames>;
   contractsByChainId: TContractsByChainId<GContractNames>;
 };
-
 /**
  * Creates a default state for the contracts context
  * @returns
